feat: drop v10 & v11 support
