+++ Resources/Private/Partials/FocuspointWizardInput.html
@@ -44,7 +44,7 @@
                         <button class="btn btn-default t3js-form-field-inputlink-explanation-toggle"
                                 type="button"
                                 title="Linkerläuterung umschalten">
-                            <core:icon identifier="{f:if(condition:'{typo3Version}>9', then:'actions-eye-link', else:'actions-version-workspaces-preview-link')}" size="small"/>
+                            <core:icon identifier="actions-eye-link" size="small"/>
                         </button>
                         <input type="hidden"
                                name="data[tt_content][4267][header_link]"
