+++ Resources/Private/Templates/FocuspointElement.html
@@ -52,11 +52,6 @@
             </f:if>
 
         </div>
-        <div class="form-wizards-items-aside">
-            <div class="btn-group">
-                {fieldControl -> f:format.raw()}
-            </div>
-        </div>
         <div class="form-wizards-items-bottom">
             {fieldWizard -> f:format.raw()}
         </div>
