+++ composer.json
@@ -34,7 +34,7 @@
   ],
   "require": {
     "ext-json": "*",
-    "typo3/cms-core": "^10.0 || ^11.0 || ^12.0"
+    "typo3/cms-core": "^12.0"
   },
   "config": {
     "allow-plugins": {
